fix(migration): change the cost column type

+++ database/migrations/2022_07_02_135232_create_messages_table.php
@@ -21,7 +21,7 @@ class CreateMessagesTable extends Migration
             $table->string('receptor');
             $table->string('message');
             $table->tinyInteger('status')->default(Message::STATUS['in_queue']);
-            $table->tinyInteger('cost')->default(0);
+            $table->smallInteger('cost')->default(0);
             $table->dateTime('publish_time')->default(now());
             $table->timestamps();
